Split GUI app shell helpers
